frontend: helpers: Extract table rows stuff to tablesRowsPerPage
