+++ frontend/src/components/App/Settings/NumRowsInput.tsx
@@ -14,7 +14,7 @@ import {
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
-import helpers from '../../../helpers';
+import { getTablesRowsPerPage, setTablesRowsPerPage } from '../../../helpers/tablesRowsPerPage';
 import { defaultTableRowsPerPageOptions, setAppSettings } from '../../../redux/configSlice';
 
 export default function NumRowsInput(props: { defaultValue: number[] }) {
@@ -28,7 +28,7 @@ export default function NumRowsInput(props: { defaultValue: number[] }) {
     }
   }, []);
   const defaultRowsPerPageValue = useMemo(() => {
-    const val = helpers.getTablesRowsPerPage();
+    const val = getTablesRowsPerPage();
     if (options.includes(val)) {
       return val;
     }
@@ -54,7 +54,7 @@ export default function NumRowsInput(props: { defaultValue: number[] }) {
   // Make sure we update the value in the localStorage when the user selects a new value.
   useEffect(() => {
     if (selectedValue !== -1) {
-      helpers.setTablesRowsPerPage(selectedValue);
+      setTablesRowsPerPage(selectedValue);
     }
   }, [selectedValue]);
